// src/app/components/dashboard/types.ts
export interface Profissional {
    id: number;
    nome: string;
    cpf: string;
    rg: string;
    nascimento: string;
    email: string;
    telefone: string;
    cidade: string;
    cep: string;
    rua: string;
    casa: string;
    bairro: string;
    cfm: string;
    especialidade: string;
}